Add a bunch of session tests
